Make the exporter resource a more explicit structure and add additiona links
Previously the primary key of the object this data is in the context of
was prepended onto the row, but had not associated key added. Each
'row' is now an object that with primary key and `values` array with
the data corresponding to the columns from the DataView instance.

The root model verbose names have been added to the exporter and
datacontext resources for additional information.