Resource overhaul, integrate preserialize posthook and restlib2 Parametizer
This significantly cleans up the code and makes it easier to extend in
the future. A few details:

- The classmethod decorator has been removed from prepare methods
- The `parametizer` attribute can be set to customize which Parametizer
class to use in `get_params`

The ConceptResource now, by default, does not embed related field within
the response. The `embed` GET param can be set to `1` to explicitly embed
the field data.

ConceptResource and FieldResource now support the `limit` and `brief`
parameters. The former limit the number of items returned and the latter
returns a lightweight response suitable for search-result based queries
that only need minimal descriptive data.