Improve integration of default template
The BaseResource get_context and get_view helper methods now
default to the default template is one is defined. Note this does
save an instance for requesting agent at this stage since these
are outside of the View- and Context- Resource classes.

Note, this now requires Avocado 2.0.23 to remove redundant
definition of get_default_template.