Don't expect a response if scipy isn't present
If scipy is not installed, then the endpoint tested in test_dist will not be
defined. Because of the variable existence of the endpoint, the test has been
updated to expect a response only when scipy is available. For more info on
the endpoint definition, see: https://github.com/cbmi/serrano/blob/2.x/serrano/resources/field/__init__.py#L20-L27.