Add POST support for validation of field values
This takes a JSON-encoded object or array and echos back the data with
the value `label` and a `valid` flag for each input.

Currently, this only checks if the value exists in the data. Future
enhancements could provide a reason why validation failed, e.g. 'does
not exist', 'out of bounds', etc.