Catch ValueError in ObjectRevisionResource is_not_found()
This should really never happen if urls are constructed correctly but we
can account for the possibility anyway. The ValueErrors from
get_object() are now caught and handled in is_not_found(). Also, tests
have been added to simulate the case where the object id and revision id
are left out.