Make shared query endpoint data distinct
The use of the OR'd Q expression was resulting in duplicate results when a user was the owner and also in the list of shared_users. In practice, this shouldn't really happen but since it is possible to be the owner and share the query with yourself, it is better to just make the results distinct to avoid any confusion.