Expose distribution endpoint for auto and key-based fields
This change also resolves a side-effect seen when calling stats_capable
which relied on the `internal_type` of the field. When a field is orphaned,
this threw an error since the underlying model field does not exist.
The tests were updated to remove handling this side effect and actually
test the result.

Fix #105