Add ping resource and endpoint for session and system status
In practice, this requires the SESSION_SAVE_EVERY_REQUEST Django setting
to be false since the act of pinging the server would refresh the session.

This also only works when Serrano requires authentication and the
AUTH_REQUIRED setting is set to true. Django's session behavior is to
transparently generate a new session key for requests if one is not
present or it has expired for the requester.

Fix #86