Execute application stat counts in parallel
A thread pool is started to get/execute the counts in parallel. The data
will be returned regardless if a count times out and will be set to
`None`. Clients can assume that the count is currently unavailable when
this occurs. The `STATS_COUNT_TIMEOUT` setting determines how long to
wait until the thread times out.

This requires the test database to not be SQLite, since multiple
connections are not supported while the test suite is running. See
https://github.com/django/django/blob/stable/1.6.x/django/db/backends/sqlite3/base.py#L96

Fix #252

Signed-off-by: Byron Ruth <b@devel.io>