Add endpoints for SQL representation of queries
This change introduces three new endpoints that returns an SQL
representation for with respect to the context, view, and query.

For example, going to /api/contexts/session/sql/ will return the SQL
for just the context whereas /api/views/session/sql/ will return the
SQL for just the view component. /api/views/queries/sql/ will return
the composite of the two which is the basis for the query results.

Signed-off-by: Byron Ruth <b@devel.io>